project_root/requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
blis==1.3.3
cachetools==6.2.6
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
distro==1.9.0
easyocr==1.7.2
emoji==2.15.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
faiss-cpu==1.13.2
filelock==3.20.3
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.15
narwhals==2.15.0
neo4j==6.1.0
networkx==3.6.1
ninja==1.13.0
nltk==3.9.2
numpy==2.4.1
opencv-python-headless==4.13.0.90
packaging==26.0
pandas==2.3.3
pdf2image==1.17.0
pillow==12.1.0
preshed==3.0.12
protobuf==6.33.5
pyarrow==23.0.0
pyclipper==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stanza==1.11.0
streamlit==1.53.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2026.1.28
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.1
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
